camera: Mark as moved when distance changes
When camera_update() changes camera's distance to character, it's not
marked as moved. The only reason it actually gets repositioned is
because the character is always in motion just enough for
character_update() to trigger camera update. The fact that the
character's body is in motion is not a given and shouldn't be relied
upon for this.

Explicitly mark camera as moved when its distance changes.

Signed-off-by: Alexander Shishkin <alex@ash.works>