pipeline-builder: Make menu blur not explode with HDR enabled
Menu blur uses the same shaders as the HDR/LDR bloom chain, but uses
LDR textures, so when HDR is enabled, the last stage drives everything
outside of the LDR color range. Fix this by replacing the last shader
with "contrast", because I'm too lazy to add a "passthrough" shader.

Signed-off-by: Alexander Shishkin <alex@ash.works>