cmake: Build ucpio as an ExternalProject
The problem with cross-compilation in cmake is that it doesn't really
differentiate between a host compiler and a target compiler, there's
just "compiler". So, in case of cross-compilation, tools that run at
built-time need to be built with a different instance of cmake, the one
that uses host compiler instead of what the main project uses via
toolchain files and preset trickery. Build ucpio as an ExternalProject
so as to do it with a separate instance of cmake to get an executable
that runs on the host at build time.

Signed-off-by: Alexander Shishkin <alex@ash.works>