compat: Move fls() and ffs() to compat/shared
Different C libraries do or don't implement fls() and ffs(). Move clap's
implementation of fls() to compat/shared and add a matching ffs(), in
case the C library doesn't have it.

This starts a compat library that's intended to cover for various
platforms' deficiencies or quirks so that the core code doesn't have to
care about them.

Signed-off-by: Alexander Shishkin <alex@ash.works>