model: Add model3dtx::shadow
Currently, model3dtx supports 4 textures, which, especially for
postprocessing quads is not enough any more. The proper fix for the
future is to maintain an array of sampler uniforms indexed by enum
shader_vars (or a subset of it). For now, just add model3dtx::shadow for
the 5th texture.

Signed-off-by: Alexander Shishkin <alex@ash.works>