pipeline: Add RM_PLUG method to use textures as sources
Currently, only render passes' framebuffers or model queues can be used
as sources for render passes. It is useful, however, to be able to use
textures that don't come out of other render passes as sources too.

Allow textures to be used directly as sources for render passes by means
of RM_PLUG.

Signed-off-by: Alexander Shishkin <alex@ash.works>