model: Find the largest bounding volume for camera/character
For shadow frustum length, it is useful to know the entity with the
largest AABB volume that contains the camera or the character. Calculate
it during scene_update() loop that already iterates over all entities.
Afterwards, clear it in scene_camera_calc() for the next frame to start
from scratch.

Signed-off-by: Alexander Shishkin <alex@ash.works>