view: Make light frusta "look_at" the centers of camera's frusta
Right now, the light frusta's view matrix is calculated to point from
the center of the coordinate system in the direction of the light
source, which results in skewed shadow maps. Instead, make it look at
the center of the light frusta from one unit behind, to fix the rotation
and remove translation from the initial view matrix.

This fixes the broken shadows at grazing light angles.

While at it, change the order of arguments of view_update_from_frustum()
to move the direction vector after struct view objects.

Signed-off-by: Alexander Shishkin <alex@ash.works>
Fixes: #127