ldjam57: Don't rebuild the pipeline needlessly
Edge antialiasing is enabled by default, but the executable assumes
otherwise, so when it finds it enabled, it triggers a pipeline rebuild,
which is completely unnecessary. Adjust its expectations accordingly.

This is necessary while the pipeline-rebuilding render options are
tracked outside of the engine as well as the rebuilding itself. This
should be done in the clap's frame function instead, but for now, fix
the immediate problem.

Signed-off-by: Alexander Shishkin <alex@ash.works>