shaders: Add Variance Shadow Mapping (VSM) shaders
Variance Shadow Maps store moments (Z and Z^2) in a color attachment, so
that the shadow sampling can use variance and Chebyshev upper bounds to
determine if a fragment is occluded. The shadow_vsm.frag is the only
thing different from the regular CSM shadow shaders. Add VSM vertex,
fragment and geometry (for Core GL) shaders and add them to the build
system.

Signed-off-by: Alexander Shishkin <alex@ash.works>