demo: Trigger pipeline rebuild on VSM toggle
Like many other render options, switching VSM shadows on or off should
trigger a pipeline rebuild.

This is necessary while the pipeline-rebuilding render options are
tracked outside of the engine as well as the rebuilding itself. This
should be done in the clap's frame function instead, but for now do it
in the demo executables' frame callbacks.

Signed-off-by: Alexander Shishkin <alex@ash.works>