clap: Store OS info in clap_context
For any OS-related queries, store the basic OS information in the clap
context and add an API to retrieve it. Specifically, obtain browser's
user-agent string in web builds and set the "mobile" property if iOS or
iPadOS is detected. The native build should use uname(2) where it's
available, but this is not done yet.

Signed-off-by: Alexander Shishkin <alex@ash.works>