ui: Fix menu item animations
If a menu item gets selected while the menu toggle animation is still
playing, item's "on focus" animation briefly kicks in making that item
appear to hop to its target position and back as the menu toggle
animation continues to play.

Fix this by cancelling menu toggle animation for that item and queuing
the "on focus" animation instead.

Signed-off-by: Alexander Shishkin <alex@ash.works>