ui-imgui: Add ImGui wrappers with label on the left side
ImGui has all widget labels on the right side, which is contrary to all
UI libraries ever created and also has some unintentional resizing
effects. Add helpers for arranging widgets in a 2-column table with
labels on the left side (except for checkboxes) and the actual widgets
on the right.

Signed-off-by: Alexander Shishkin <alex@ash.works>