util: Convert hashmap_insert() to cerr
Currently, hashmap_insert() return -EBUSY or -ENOMEM or 0, which are not
check for (by the sole user). Convert it to return cerr errors or
success. This temporarily introduces an unchecked return value warning
in test.c, which is fine because it's not built with -Werror (yet) and
will be fixed in upcoming commits.

Signed-off-by: Alexander Shishkin <alex@ash.works>