mesh: Fix a subtle meshopt_simplify() misuse
meshopt_simplify*() functions return size_t number of edges, accept
size_t as the target number of edges and don't use pre-initialized
result error. Correct these by converting edge counts to size_t, not
checking the returned result from the simplify functions for being
negative and initializing result_error to 0.0. The last one is strictly
speaking unnecessary: as of today, meshopt_simplifyEdge() always writes
it if the pointer is non-NULL, but if this changes one day this way we'd
be covered.

Signed-off-by: Alexander Shishkin <alex@ash.works>