model: Move current LOD from model3d to entity3d
Obviously, current LOD is a property of an entity, not a model. This
requires that index buffer binding needs to move out of model3d_prepare()
and the same goes for the unbinding and re-binding when the LOD changes
from entity to entity. Make them open-coded in models_render() for now.

While at it, fix the LOD calculation to take into account entity's AABB,
distance from its center to the camera and its average edge.

Signed-off-by: Alexander Shishkin <alex@ash.works>