core: Add LUT management infrastructure
To facilitate color grading the final rendered image, add support for
LUT textures. They can be loaded from .cube files or generated at
runtime from a number of presets, and also switched between at runtime.

To start things off, there are 12 presets covering a range of color
grading purposes, so that no additional assets need to be
included/shipped with the resulting executable. Additionally, a LUT can
be loaded from a .cube file, although this loader is limited to 3D
arrays of sRGB ranged values.

A LUT (lookup table) is a 3D texture that maps RGB values to R'G'B'
values (with linear interpolation in between specified values) and is
applied towards the end of the rendering pipeline to the final composite
image. Separate LUTs can be used for different things (lighting pass,
bloom pass, characters, materials) if the user wishes.

Signed-off-by: Alexander Shishkin <alex@ash.works>