camera: Convert to transform_t
Currently, camera uses a character (which brings with it entity3d,
model3dtx and model3d) for position and "updated" status, which is
wildly unnecessary and wasteful. Convert it to use transform_t, which
provides all the same benefits without all the overhead.

The one downside is that is that it can no longer be controlled as a
"detached" camera using character's motion code. But even that required
a bunch of special treatment for "character is camera" case.

Signed-off-by: Alexander Shishkin <alex@ash.works>