shaders/model: Fix TBN calculation
For models with normal maps and tangents, the TBN vectors are calculated
in view space, but disregard entity's rotation. Also, normal vectors
from the normal map are in world space. Fix this by applying entity's
rotation and view rotatio in TBN vector calculation and passing the tbn
matrix to the fragment shader, so that normal map's sampled vectors are
translated into TBN space.

Signed-off-by: Alexander Shishkin <alex@ash.works>