scene: Add model selector to entity picker
Some of the properties in the entity picker belong to the model and not
the entity. Also, given that models are designed to have multiple
entities, mashing them together is not optimal.

Add a top-level model picker, so that the entity picker can then only
select entities of that model.

While at it, add a checkbox that determines whether the selected entity
automatically becomes "controlled" by the scene (camera focus, motion
control move the entity etc) and a checkbox that makes the entity picker
follow the currently controlled entity.

Signed-off-by: Alexander Shishkin <alex@ash.works>