www: Handle resize event the same way as display-glfw.c
The resize event in native builds call a callback (clap_resize()), which
then handles the internal data structures, including
scene::{width,height}, but input-www.c instead sends a "resize" input
event, which is captured by scene_handle_input(), which in turn calls
display_resize(), which ends up in clap_resize() as well. This is a
legacy from the very first days of clap and is totally backwards.

Instead, do the same in input-www.c and call clap_resize() directly.

Signed-off-by: Alexander Shishkin <alex@ash.works>