util: Fix UBSAN warnings in darray_for_each()
When called on an empty darray, darray_for_each() technically
dereferences a NULL array in 2 places, which in practice cancel each
other out and don't create a problem. But a cleaner way is not to do
that. While technically not a bug, it has potential to become one. Fix
it by using an element accessor instead of plain dereferencing and
guarding the other one with a NULL check.

Signed-off-by: Alexander Shishkin <alex@ash.works>