cmake, shaders: Convert to Vulkan flavor of GLSL
Currently, glslc requires --target-env=opengl to compile the shaders,
because Vulkan is more strict and in some ways incompatible with OpenGL
shaders. Now, all the Vulkan requirements have been addressed, except
the gl_InstanceID variable does not conform, but for our purposes,
gl_InstanceIndex can be a direct drop-in replacement.

Convert gl_InstanceID and remove --target-env=opengl, so that the
shaders are fully compatible with Vulkan and still work perfectly fine
with GL.

Signed-off-by: Alexander Shishkin <alex@ash.works>