shader(s): Share attribute locations between CPU side and GPU side
Currently, we're relying on attribute locations in the shaders matching
attribute constants in enum shader_vars, and the only reason this
doesn't explode is the superfluous glVertexAttribPointer() that gets
called on *every* buffer bind, which is both wasteful and masking the
problem. Make the shaders use attribute location constants and set
corresponding shader_vars to the same constants.

Signed-off-by: Alexander Shishkin <alex@ash.works>