render-gl: Remove superfluous glVertexAttribPointer()
Currently, buffer_bind() calls glVertexAttribPointer() every frame, for
every attribute of every model, which is completely unnecessary and in
case of WebGL, wasteful. Now that the shaders share attribute locations
with the core shader code, the location given at buffer_init() time does
not change, so there's no need to re-affirm it. Remove the redundant
call.

Signed-off-by: Alexander Shishkin <alex@ash.works>