mesh: Add mesh_flatten()
Add a function that creates a contiguous buffer with all of the mesh
attributes interleaved. Given an array of mesh attributes, element
sizes, offsets within a vertex stride and the length of a stride,
mesh_flatten() combines these attributes into a single contiguous buffer
that can then be bound to a single binding point.

Signed-off-by: Alexander Shishkin <alex@ash.works>