object: Use REFCLASS_NAME() instead of its open-coded version
In a few places, instead of REFCLASS_NAME() the code spells out the same
thing. Replace those with REFCLASS_NAME(). It doesn't improve the
readability of some of the more involved macros there, but at least
they'll be using the single source of truth, which would make
refactoring a little easier, if need be.

Signed-off-by: Alexander Shishkin <alex@ash.works>