shaders: Calculate to_light_vectors in the fragment shader
Currently, light direction vectors are calculated in model's vertex
shader and passed as an array to the fragment shader, which saves a few
cycles (calculating per-vertex instead of per-fragment), but has an
unfortunate problem of exhausting the available layout slots as the
number of light sources grows.

Move the calculation to lighting.glsl, which has all the necessary
information to perform it.

Signed-off-by: Alexander Shishkin <alex@ash.works>