shaders/model: Get rid of "normal_vec"
In model.frag, normal_vec doesn't serve any special purpose other than
being a transient variable, yet somehow it's fed into the SSAO shader,
which, in the even of there being a model with a normal map, may not
look great. Remove normal_vec and replace it with surface normal
orig_normal as a source for SSAO shader.

Signed-off-by: Alexander Shishkin <alex@ash.works>