shader: Expliticly bind texture slot to its uniform
In GLSL ES, there are no layouts, so there are no locations assigned to
the opaque uniforms that spirv-cross produces for "--es --version 300".
The texture uniforms that are not explicitly set to slots end up
assigned slots in the order of appearance, which may go slightly
backwards if one of the included files declares a texture uniform before
the main TU of the fragment shader.

Long story short, fix shader_plug_texture() to always do this regardless
of whether there is a texture to bind or not.

Signed-off-by: Alexander Shishkin <alex@ash.works>