model: Initialize mq::nr_characters
Currently, mq_init() forgets to initialize its nr_characters. For the
scene object this works out well, because it's embedded there and
scene_init() memset()s itself to 0 before anything else. For the ui
object it also works out, because it's embedded in clap context, which
is zero initialized on allocation.

However, it's bad for and a source of potentially nasty future bugs.
Initialize it explicitly.

Signed-off-by: Alexander Shishkin <alex@ash.works>