render-gl: Fix texture loading
Currently, texture_load() and texture_fbo() are trying to be clever and
bypass format/internal_format/component_type updates if incoming format
is the same as texture's format. But this is OpenGL and the format alone
doesn't uniquely identify a color format. For example RGBA8 and RGBA32F
have the same format GL_RGBA, so if a texture was initialized as
TEX_FMT_RGBA8 (which is the default), trying to texture_load() a
TEX_FMT_RGBA32F buffer will not set its format to RGBA32F, providing for
all manner of hilarity of varying degree of subtlety. In other words,
the microoptimization that skips the format update is, in fact, a bug.

Fix this by always updating texture's format in texture_load() and
texture_fbo().

While at it, fix a related microoptimization texture_t::updated, which,
if false, skips redundant sampler configuration calls, but inexplicably
ends up set to true on texture resize, rendering itself practically
useless. Rename it to texture_t::sampler_updated and only set it at
texture initialization time, since that's the only way to set up
texture's sampler parameters.

Signed-off-by: Alexander Shishkin <alex@ash.works>