sound: Replace OpenAL with miniaudio
Miniaudio can do pretty much everything that OpenAL 1.1 can do (which is
what we get from emscripten's emulation), plus it can decode ogg, mp3
and wav files without external dependencies (it carries a copy of
stb_vorbis.h). It's much lighter than openal-soft: just 1 translation
unit. And since at the moment the spatial features of OpenAL are not
even used, it's basically a glorified PCM player, there is no reason to
keep it around, and all the reasons listed above to replace it with
miniaudio.

There are some caveats:
- openal-soft has built-in effects, miniaudio does not, although it has
  all the APIs necessary to implement them, in which case they'll work
  in wasm builds as well, unlike with emscripten's openal implementation
- miniaudio requires web workers and shared memory, which browsers only
  allow for https served pages with no cross-origin, so all fonts etc
  have to be served from the same origin and python3 -m http.server
  oneliner won't work any more. This will be addressed in upcoming
  commits.

Signed-off-by: Alexander Shishkin <alex@ash.works>