cmake, shaders: Add CONFIG_SHADOW_MAP_ARRAY
Currently, there are more than a few instances where CONFIG_GLES is used
to determine whether shadow maps are a texture array or multiple
separate textures. This is not really exclusive to GLES, but anywhere
where there aren't geometry shaders.

Add CONFIG_SHADOW_MAP_ARRAY to use in the code and shaders for this
purpose instead of CONFIG_GLES, replace all relevant instances of the
latter with the former.

Signed-off-by: Alexander Shishkin <alex@ash.works>