model: Clean up and rename load_gl_texture_buffer()
First, it's not GL specific, so having a _gl_ infix is unnecessary, so
drop it. Second, instead of has_alpha, make it take texture_format as a
parameter and let the callers decide what it is, in case the incoming
buffer is (half-)float or any component width larger than 8-bits. Third,
fix coding style. And finally, change the caller to pass texture_format
instead of has_alpha. No functional changes.

Signed-off-by: Alexander Shishkin <alex@ash.works>