clap: Disable ASAN string checks on macOS + OpenGL
There's a known old bug in macOS where it trips up ASAN trying to obtain
its mouse cursors from parsing a PDF (!) [1]. This used to happen
occasionally on mouse hover over the window's edge or corner, but with
the advent of macOS 26 it happens 100% of the time on the initial window
size set (glfwSetWindowSize()) before there's even a window to hover
over. Funnily, this only happens with the OpenGL renderer, not Metal.

So, for this combination, disable ASAN's strict_string_checks, to
sidestep this.

[1] https://bugzilla.mozilla.org/show_bug.cgi?id=1593160

Signed-off-by: Alexander Shishkin <alex@ash.works>