util: Make aabb_center() a static inline
GCC's LTO gets triggered at one of aabb_center()'s call sites,
incorrectly inferring that the input aabb is vec3 instead of vec3[2]. It
is counterproductive to change code to pacify compilers' false
positives, but the function is short enough to be a static inline, which
would move the stringop-overflow check burden onto the compiler, which
instead does it correctly and makes the linker warning naturally go
away. So, make it static inline.

Signed-off-by: Alexander Shishkin <alex@ash.works>