core: Move messagebus into clap_context
Currently, messagebus is a global static in messagebus.c, which creates
a number of problems. Moving it to clap_context is a good longterm
solution, but it requires clap_context pointer be passed into all
messagebus APIs, and as a result, all their users, which is a very
invasive change, but again, a necessary one. Do it.

Signed-off-by: Alexander Shishkin <alex@ash.works>