render-gl: Make every GL call validation switchable
In debug GL builds, every GL call is validated by glGetError() to catch
invalid behavior early. This is a landmine, however, for many browsers
that slows effectively doubles the number of expensive WebGL calls.

Instead of calling it unconditionally, add an off-by-default checkbox to
enable this, making debug builds usable in browsers.

Signed-off-by: Alexander Shishkin <alex@ash.works>